Más mejoras, paso 5

+++ data_param/api/views.py
@@ -1,23 +1,32 @@
-# -*- coding: utf-8 -*-
 from rest_framework.decorators import action
 from rest_framework.exceptions import PermissionDenied
+from rest_framework.pagination import PageNumberPagination
 
-import data_param.api.serializers
-import inai.api.serializers
+
 from api.mixins import MultiSerializerModelViewSet
-from inai.api import serializers
+from inai.api.serializers import (
+    PetitionSemiFullSerializer, PetitionFileControlDeepSerializer,
+    FileControlFullSerializer, TransformationEditSerializer,
+    NameColumnEditSerializer)
 from inai.models import PetitionFileControl, DataFile
 from task.views import build_task_params, comprobate_status
 from . import serializers
 from rest_framework.response import Response
-from rest_framework import (permissions, views, status)
+from rest_framework import (permissions, status)
 
 from ..models import FileControl, Transformation, NameColumn
+
+
+class HeavyResultsSetPagination(PageNumberPagination):
+    page_size = 20
+    page_size_query_param = 'page_size'
+    max_page_size = 100
 
 
 class FileControlViewSet(MultiSerializerModelViewSet):
     permission_classes = (permissions.AllowAny,)
     serializer_class = serializers.FileControlSerializer
+    pagination_class = HeavyResultsSetPagination
     queryset = FileControl.objects.all().prefetch_related(
         "data_group",
         "columns",
@@ -36,20 +45,36 @@ class FileControlViewSet(MultiSerializerModelViewSet):
         return context
 
     action_serializers = {
-        "list": inai.api.serializers.FileControlFullSerializer,
-        "retrieve": inai.api.serializers.FileControlFullSerializer,
+        "list": FileControlFullSerializer,
+        "retrieve": FileControlFullSerializer,
         "create": serializers.FileControlSerializer,
         "post": serializers.FileControlSerializer,
         "update": serializers.FileControlSerializer,
+        "filter": serializers.FileControlSemiFullSerializer,
     }
+
+    def get_queryset(self):
+        request = self.request
+        status_register = request.query_params.get('status_register', None)
+        controls = FileControl.objects.all().prefetch_related(
+            "data_group",
+            "columns",
+            "columns__column_transformations",
+            "file_transformations",
+            "petition_file_control",
+            "petition_file_control__data_files",
+            # "petition_file_control__data_files__origin_file",
+        )
+        if status_register:
+            controls = controls.filter(status_register_id=status_register)
+        return controls
 
     def get(self, request):
         # print("ESTOY EN GET")
         file_control = self.get_object()
-        serializer = inai.api.serializers.FileControlFullSerializer(
+        serializer = FileControlFullSerializer(
             file_control, context={ 'request': request })
         return Response(serializer.data, status=status.HTTP_200_OK)
-        return Response()
 
     def create(self, request, **kwargs):
         # print("ESTOY EN CREATE")
@@ -70,7 +95,7 @@ class FileControlViewSet(MultiSerializerModelViewSet):
             new_pet_file_ctrl = PetitionFileControl.objects.create(
                 petition_id=petition_id, file_control=file_control)
 
-            new_serializer = serializers.PetitionFileControlDeepSerializer(
+            new_serializer = PetitionFileControlDeepSerializer(
                 new_pet_file_ctrl, context={ 'request': request })
             return Response(
                 new_serializer.data, status=status.HTTP_201_CREATED)
@@ -79,12 +104,10 @@ class FileControlViewSet(MultiSerializerModelViewSet):
                 {"errors": "No se ha podido crear el control de archivo"},
                 status=status.HTTP_400_BAD_REQUEST)
 
-
     def update(self, request, **kwargs):
         file_control = self.get_object()
         data = request.data
 
-        transformations = data.pop('transformations', [])
         serializer_file_control = self.get_serializer_class()(
             file_control, data=data)
         if serializer_file_control.is_valid():
@@ -93,19 +116,21 @@ class FileControlViewSet(MultiSerializerModelViewSet):
             return Response({ "errors": serializer_file_control.errors },
                             status=status.HTTP_400_BAD_REQUEST)
 
-        actual_id_tranformations = []
-        for transf_item in transformations:
-            transformation = Transformation()
-            transform_ser = serializers.TransformationEditSerializer(
-                transformation, data=transf_item)
-            if transform_ser.is_valid():
-                tranform = transform_ser.save()
-                actual_id_tranformations.append(tranform.id)
-        Transformation.objects.filter(file_control=file_control) \
-            .exclude(id__in=actual_id_tranformations).delete()
+        transformations = data.pop('transformations', None)
+        if transformations is not None:
+            actual_id_tranformations = []
+            for transf_item in transformations:
+                transformation = Transformation()
+                transform_ser = TransformationEditSerializer(
+                    transformation, data=transf_item)
+                if transform_ser.is_valid():
+                    tranform = transform_ser.save()
+                    actual_id_tranformations.append(tranform.id)
+            Transformation.objects.filter(file_control=file_control) \
+                .exclude(id__in=actual_id_tranformations).delete()
 
         new_file_control = FileControl.objects.get(id=file_control.id)
-        new_serializer = inai.api.serializers.FileControlFullSerializer(
+        new_serializer = FileControlFullSerializer(
             new_file_control)
         return Response(
             new_serializer.data, status=status.HTTP_206_PARTIAL_CONTENT)
@@ -113,10 +138,53 @@ class FileControlViewSet(MultiSerializerModelViewSet):
         # return Response(
         #    serializer_file_control.data, status=status.HTTP_206_PARTIAL_CONTENT)
 
+    @action(methods=["get"], detail=False, url_path='filter')
+    def filter(self, request, **kwargs):
+        from data_param.models import FileControl
+        from inai.models import Petition
+        import json
+        limiters = request.query_params.get("limiters", None)
+        limiters = json.loads(limiters)
+        controls = FileControl.objects.all().prefetch_related(
+            "data_group",
+            "columns",
+            "columns__column_transformations",
+            "entity",
+        ).order_by("data_group__name", "entity__acronym", "name")
+        total_count = 0
+        if limiters:
+            if limiters.get("status_register"):
+                controls = controls.filter(
+                    status_register_id=limiters.get("status_register"))
+            if limiters.get("data_group"):
+                controls = controls.filter(
+                    data_group_id=limiters.get("data_group"))
+            if limiters.get("has_notes"):
+                controls = controls.filter(
+                    notes__isnull=not limiters.get("has_notes"))
+            total_count = controls.count()
+            page_size = limiters.get("page_size", 40)
+            page = limiters.get("page", 1) - 1
+            controls = controls[page * page_size:(page + 1) * page_size]
+        if not total_count:
+            total_count = controls.count()
+        # serializer =
+        serializer = serializers.FileControlSemiFullSerializer(
+            controls, many=True, context={ 'request': request })
+        related_petitions = Petition.objects.filter(
+            file_controls__file_control__in=controls).distinct()
+        serializer_petitions = PetitionSemiFullSerializer(
+            related_petitions, many=True, context={ 'request': request })
+        data = {
+            "file_controls": serializer.data,
+            "petitions": serializer_petitions.data,
+            "total_count": total_count,
+        }
+        return Response(data, status=status.HTTP_200_OK)
+
     @action(methods=["post"], detail=True, url_path='columns')
     def columns(self, request, **kwargs):
         from inai.models import FileControl
-        import json
         if not request.user.is_staff:
             raise PermissionDenied()
         columns_items = request.data.get("columns")
@@ -144,7 +212,7 @@ class FileControlViewSet(MultiSerializerModelViewSet):
                 column = NameColumn()
                 column.file_control = file_control
 
-            column_serializer = serializers.NameColumnEditSerializer(
+            column_serializer = NameColumnEditSerializer(
                 column, data=column_item)
             if column_serializer.is_valid():
                 print("es válido")
@@ -154,7 +222,7 @@ class FileControlViewSet(MultiSerializerModelViewSet):
                 for transf_item in transformations:
                     transformation = Transformation()
                     transf_item["name_column"] = column.id
-                    transform_ser = serializers.TransformationEditSerializer(
+                    transform_ser = TransformationEditSerializer(
                         transformation, data=transf_item)
                     if transform_ser.is_valid():
                         print("transf válido")
@@ -172,7 +240,7 @@ class FileControlViewSet(MultiSerializerModelViewSet):
             .exclude(id__in=actual_id_columns).delete()
 
         new_file_control = FileControl.objects.get(id=file_control.id)
-        new_serializer = inai.api.serializers.FileControlFullSerializer(
+        new_serializer = FileControlFullSerializer(
             new_file_control)
         return Response(
             new_serializer.data, status=status.HTTP_201_CREATED)
@@ -214,8 +282,10 @@ class FileControlViewSet(MultiSerializerModelViewSet):
             status_process_id=status_id)
         method = status_req.get("addl_params", { }).get("next_step", { }).get("method", False)
         if method == "buildExploreDataFile" or method == "countFile":
+            function_name = "massive_explore" \
+                if method == "buildExploreDataFile" else "massive_count"
             key_task, task_params = build_task_params(
-                file_control, "massive_explore", request, subgroup=status_name)
+                file_control, function_name, request, subgroup=status_name)
             all_tasks = []
             all_errors = []
 
@@ -248,7 +318,7 @@ class FileControlViewSet(MultiSerializerModelViewSet):
 
             data = {
                 "errors": all_errors,
-                "file_control": inai.api.serializers.FileControlFullSerializer(
+                "file_control": FileControlFullSerializer(
                     file_control).data,
             }
             comprobate_status(key_task, errors=all_errors, new_tasks=all_tasks)
